requirements.txt
langchain
langchain-google-genai
langchain-groq
langchain-openai
langchain-community
streamlit
python-dotenv
fpdf2
openpyxl
pandas
matplotlib
seaborn
python-pptx
youtube-transcript-api
pytube
moviepy
openai-whisper
requests
beautifulsoup4
